refactor by review
